requirements.txt
geopandas
matplotlib
pandas
pillow
plotly
plotly_express
shapely
streamlit

